split eid tests into each userId submodule
