+++ test/spec/modules/33acrossIdSystem_spec.js
@@ -3,6 +3,9 @@ import * as utils from 'src/utils.js';
 
 import { server } from 'test/mocks/xhr.js';
 import { uspDataHandler, coppaDataHandler, gppDataHandler } from 'src/adapterManager.js';
+import {createEidsArray} from '../../../modules/userId/eids.js';
+import {expect} from 'chai/index.mjs';
+import {attachIdSystem} from '../../../modules/userId/index.js';
 
 describe('33acrossIdSystem', () => {
   describe('name', () => {
@@ -817,4 +820,25 @@ describe('33acrossIdSystem', () => {
       });
     });
   });
+  describe('eid', () => {
+    before(() => {
+      attachIdSystem(thirthyThreeAcrossIdSubmodule);
+    })
+    it('33acrossId', function() {
+      const userId = {
+        '33acrossId': {
+          envelope: 'some-random-id-value'
+        }
+      };
+      const newEids = createEidsArray(userId);
+      expect(newEids.length).to.equal(1);
+      expect(newEids[0]).to.deep.equal({
+        source: '33across.com',
+        uids: [{
+          id: 'some-random-id-value',
+          atype: 1
+        }]
+      });
+    });
+  })
 });
