+++ test/spec/modules/ftrackIdSystem_spec.js
@@ -3,10 +3,10 @@ import * as utils from 'src/utils.js';
 import { uspDataHandler } from 'src/adapterManager.js';
 import { loadExternalScript } from 'src/adloader.js';
 import { getGlobal } from 'src/prebidGlobal.js';
-import { init, setSubmoduleRegistry } from 'modules/userId/index.js';
+import {attachIdSystem, init, setSubmoduleRegistry} from 'modules/userId/index.js';
 import {createEidsArray} from 'modules/userId/eids.js';
 import {config} from 'src/config.js';
-let expect = require('chai').expect;
+import 'src/prebid.js';
 
 let server;
 
@@ -557,6 +557,41 @@ describe('FTRACK ID System', () => {
           }]);
         });
       });
+    });
+  });
+  describe('eid', () => {
+    before(() => {
+      attachIdSystem(ftrackIdSubmodule);
+    });
+    it('should return the correct EID schema', () => {
+      // This is the schema returned from the ftrack decode() method
+      expect(createEidsArray({
+        ftrackId: {
+          uid: 'test-device-id',
+          ext: {
+            DeviceID: 'test-device-id',
+            SingleDeviceID: 'test-single-device-id',
+            HHID: 'test-household-id'
+          }
+        },
+        foo: {
+          bar: 'baz'
+        },
+        lorem: {
+          ipsum: ''
+        }
+      })).to.deep.equal([{
+        source: 'flashtalking.com',
+        uids: [{
+          atype: 1,
+          id: 'test-device-id',
+          ext: {
+            DeviceID: 'test-device-id',
+            SingleDeviceID: 'test-single-device-id',
+            HHID: 'test-household-id'
+          }
+        }]
+      }]);
     });
   })
 });
