+++ test/spec/modules/lotamePanoramaIdSystem_spec.js
@@ -6,6 +6,8 @@ import { uspDataHandler } from 'src/adapterManager.js';
 import * as utils from 'src/utils.js';
 import { server } from 'test/mocks/xhr.js';
 import sinon from 'sinon';
+import {attachIdSystem} from '../../../modules/userId/index.js';
+import {createEidsArray} from '../../../modules/userId/eids.js';
 
 const responseHeader = { 'Content-Type': 'application/json' };
 
@@ -21,7 +23,6 @@ describe('LotameId', function() {
   let requestHost;
 
   const nowTimestamp = new Date().getTime();
-
   beforeEach(function () {
     logErrorStub = sinon.stub(utils, 'logError');
     getCookieStub = sinon.stub(storage, 'getCookie');
@@ -958,4 +959,20 @@ describe('LotameId', function() {
       });
     });
   });
+  describe('eid', () => {
+    before(() => {
+      attachIdSystem(lotamePanoramaIdSubmodule);
+    });
+    it('lotamePanoramaId', function () {
+      const userId = {
+        lotamePanoramaId: 'some-random-id-value',
+      };
+      const newEids = createEidsArray(userId);
+      expect(newEids.length).to.equal(1);
+      expect(newEids[0]).to.deep.equal({
+        source: 'crwdcntrl.net',
+        uids: [{ id: 'some-random-id-value', atype: 1 }],
+      });
+    });
+  })
 });
